fixed broken tests / fixed typos created in refactor

+++ test/schema_manager.coffee
@@ -9,16 +9,19 @@ RikkiTikkiAPI.getOptions = ->
 RikkiTikkiAPI.getSchemaTree = -> {}
 module.exports.RikkiTikkiAPI = RikkiTikkiAPI
 Util = RikkiTikkiAPI.Util
-describe 'RikkiTikkiAPI.SchemaManager Test Suite', ->
+describe 'SchemaManager Test Suite', ->
+  @sm = RikkiTikkiAPI.getSchemaManager()
   it 'should List existing schema', (done)=>
-    (@sm = RikkiTikkiAPI.getSchemaManager()
-    ).on( 'error', (e)->
-      throw e
-    ).on 'success', (schemas)=>
-      @sm.listSchemas (e, list)=>
-        throw e if e?
-        list.length.should.equal 1
-        done()
+    # (@sm = RikkiTikkiAPI.getSchemaManager()
+    # ).on( 'error', (e)->
+      # console.log 'err'
+      # throw e
+    # ).on 'success', (schemas)=>
+      # console.log 'success'
+    @sm.listSchemas (e, list)=>
+      throw e if e?
+      list.length.should.equal 1
+      done()
   it 'should Create a new schema', (done)=>
     @sm.createSchema 'CreatedSchema', (e,s)=>
       throw e if e?
@@ -38,10 +41,11 @@ describe 'RikkiTikkiAPI.SchemaManager Test Suite', ->
   it 'should non-destructively Destroy a schema', (done)=>
     @sm.destroySchema 'RenamedSchema', (e,s)=>
       throw e if e?
-      throw 'error: SchemaManager was destructive in non-detructve mode' unless fs.existsSync "#{RikkiTikkiAPI.SCHEMA_PATH}/_RenamedSchema.js"
+      unless fs.existsSync "#{RikkiTikkiAPI.SCHEMA_PATH}/_RenamedSchema.js"
+        throw 'error: SchemaManager was destructive in non-detructve mode'
+        fs.unlinkSync "#{RikkiTikkiAPI.SCHEMA_PATH}/_RenamedSchema.js"
       done()
   it 'should destructively Destroy a schema', (done)=>
-    fs.unlinkSync "#{RikkiTikkiAPI.SCHEMA_PATH}/_RenamedSchema.js"
     @sm = RikkiTikkiAPI.getSchemaManager()
     setTimeout (=>
       RikkiTikkiAPI.DESTRUCTIVE = true
@@ -49,12 +53,11 @@ describe 'RikkiTikkiAPI.SchemaManager Test Suite', ->
         @sm.destroySchema 'DeleteMe', (e,s)=>
           RikkiTikkiAPI.DESTRUCTIVE = false
           throw e if e?
-          throw 'error: SchemaManager was non-destructive in detructve mode' if fs.existsSync "#{RikkiTikkiAPI.SCHEMA_PATH}/_DeleteMe.js"
+          if fs.existsSync "#{RikkiTikkiAPI.SCHEMA_PATH}/_DeleteMe.js"
+            throw 'error: SchemaManager was non-destructive in detructve mode'
+            fs.unlinkSync "#{RikkiTikkiAPI.SCHEMA_PATH}/_DeleteMe.js"
           done()
-    ), 200  
-      
-      
+    ), 200
       # console.log sm.fetchSchema( schema ).toAPISchema().toSource()
       # Foo = (new RikkiTikkiAPI.model schema, sm.__loader.__schema[schema])
       # console.log (new Foo {name:'Foo'}).toClientSchema().toString()
-
